requirements.txt
# Core
python>=3.9

# OCR
pytesseract>=0.3.10
easyocr>=1.7.0
pdf2image>=1.16.3
pymupdf>=1.23.0
pdfplumber>=0.10.0
Pillow>=10.0.0
groq>=0.4.2

# NLP & ML
transformers>=4.35.0
sentence-transformers>=2.2.0
torch>=2.0.0
spacy>=3.7.0

# Vector Store
faiss-cpu>=1.7.4

# Analytics
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# LLM
google-generativeai>=0.3.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# UI
streamlit>=1.28.0

# Utils
python-dotenv>=1.0.0
tqdm>=4.66.0
